hotfix ver0.37

+++ backend/book/views.py
@@ -332,8 +332,9 @@ def delete_comment(request):
     if request.method=='POST':
         data = json.loads(request.body)
         comment_id = data.get('commentID')
+        user_id = data.get('userID')
         try:
-            Book_Comment.objects.filter(id=comment_id).delete()
+            Book_Comment.objects.filter(id=comment_id,userID=user_id).delete()
             return JsonResponse({'message':'Comment deleting successfully'})
         except (Book_Comment.DoesNotExist):
             return JsonResponse({'message': 'Comment not found'}, status=404)
