+++ backend/game/views.py
@@ -324,8 +324,9 @@ def delete_comment(request):
     if request.method=='POST':
         data = json.loads(request.body)
         comment_id = data.get('commentID')
+        user_id = data.get('userID')
         try:
-            Game_Comment.objects.filter(id=comment_id).delete()
+            Game_Comment.objects.filter(id=comment_id,userID=user_id).delete()
             return JsonResponse({'message':'Comment deleting successfully'})
         except (Game_Comment.DoesNotExist):
             return JsonResponse({'message': 'Comment not found'}, status=404)
